+++ app/Controllers/AdminController.php
@@ -84,6 +84,7 @@ class AdminController {
             $currentYear = date('Y');
             $monthlyIncome = Payment::getMonthlyIncome($currentYear, $currentMonth, $resortId);
             $bookingHistory = Booking::getBookingHistory(10); // This could also be filtered if needed
+            $dailyIncomeData = Payment::getDailyIncomeForMonth($currentYear, $currentMonth, $resortId);
 
             // Get counts for dashboard buttons
             $pendingPaymentCount = Payment::getPendingPaymentCount($resortId);